mv(cli): new file for commands
