--- src/lib.rs
@@ -1,4 +1,0 @@
-/// This module handles reading from inputs
-pub mod file;
-/// This module handles multi sources
-pub mod source;
